tests/conftest.py
"""
Pytest configuration: load variables from `.env.example` when tests run.
"""

from pathlib import Path

from dotenv import load_dotenv

_PROJECT_ROOT = Path(__file__).resolve().parent.parent
load_dotenv(_PROJECT_ROOT / ".env.example", override=False)
